pyproject.toml
[project]
name = "learn-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "duckduckgo-search>=8.0.4",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.5",
    "langchain-huggingface>=0.3.0",
    "langchain-ollama>=0.3.3",
    "numpy>=2.3.0",
    "ollama>=0.5.1",
    "pypdf>=5.6.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
]
